refactor: NameAndResult 클래스 domain -> dto으로 옮김
